Move test and use higher-level APIs
